Upload User Manual and Original Circular Filenames
Final Versions:
36 - Titles printed with repeatition
37 - Titles printed without repetion
39 - Combined codes of AJ and AG
40 - Combined codes but with titles rather than "download" as final name
41 - With original pdf file names for circular
42 - found 31 files (32 exist out)
43 - Final version with the 31 files + warnings removed 
44 - With User manual for download

Initially, update circular files were saved using the link names. We modified this to save files using the titles provided in a table. The code was formatted using a table view with column and row tags.

However, since the titles were too long for file names, we reverted to using the original PDF file names. The logic for this involves the code first checking the content disposition from the HTTP response header. If unavailable, it then uses the last part of the document URL.

We ensured that files were not duplicated during the download process, which required several iterations. In the final version (42), we successfully scraped 30 out of 31 files. We need to investigate the missing file, which is likely due to a naming issue.

Since the user manual takes too long to load, we are considering creating a downloadable PDF version. This will reduce computational costs and provide users with a convenient reference guide.
Note: The User Manual needs to be in the same directory